Adding 2.2

+++ dev/tests/integration/testsuite/Magento/AuthorizenetAcceptjs/Fixture/full_order_with_capture_rollback.php
@@ -8,6 +8,7 @@ declare(strict_types=1);
 use Magento\Framework\Api\SearchCriteriaBuilder;
 use Magento\Sales\Api\OrderRepositoryInterface;
 use Magento\TestFramework\ObjectManager;
+use Magento\Framework\Exception\NoSuchEntityException;
 
 $objectManager = ObjectManager::getInstance();
 
@@ -18,11 +19,15 @@ $searchCriteria = $searchCriteriaBuilder->addFilter('increment_id', '100000001')
 
 /** @var OrderRepositoryInterface $orderRepository */
 $orderRepository = $objectManager->get(OrderRepositoryInterface::class);
-$items = $orderRepository->getList($searchCriteria)
-    ->getItems();
 
-foreach ($items as $item) {
-    $orderRepository->delete($item);
+$items = $orderRepository->getList($searchCriteria)->getItems();
+$item = reset($items);
+
+if ($item !== false) {
+    try {
+        $orderRepository->delete($item);
+    } catch (NoSuchEntityException $e) {
+    }
 }
 
-require __DIR__ . '/../../../Magento/Catalog/_files/product_simple_rollback.php';
+require __DIR__ . '/../_files/full_order_rollback.php';
